Add heartbeat sending implementation
- Reorganises state struct to have a separate struct for volatile state
  of leaders which is stored as an Option by Node, using it to determine
  if the node is a leader or not
- Heartbeat is sent to all followers when a node is leader

+++ src/main.rs
@@ -7,7 +7,7 @@ use tracing::trace;
 
 #[derive(Parser)]
 struct Cli {
-    node_id: usize,
+    node_id: u32,
 }
 
 #[tokio::main]
